// typescript/aptos/transaction/testing1/v1/transaction_pb.d.ts
export class Transaction extends jspb.Message {
  hasTimestamp(): boolean;
  clearTimestamp(): void;
  getTimestamp(): aptos_util_timestamp_timestamp_pb.Timestamp | undefined;
  setTimestamp(value?: aptos_util_timestamp_timestamp_pb.Timestamp): void;

  getVersion(): number;
  setVersion(value: number): void;

  hasInfo(): boolean;
  clearInfo(): void;
  getInfo(): TransactionInfo | undefined;
  setInfo(value?: TransactionInfo): void;

  getEpoch(): number;
  setEpoch(value: number): void;

  getBlockHeight(): number;
  setBlockHeight(value: number): void;

  getType(): Transaction.TransactionTypeMap[keyof Transaction.TransactionTypeMap];
  setType(value: Transaction.TransactionTypeMap[keyof Transaction.TransactionTypeMap]): void;

  hasBlockMetadata(): boolean;
  clearBlockMetadata(): void;
  getBlockMetadata(): BlockMetadataTransaction | undefined;
  setBlockMetadata(value?: BlockMetadataTransaction): void;

  hasGenesis(): boolean;
  clearGenesis(): void;
  getGenesis(): GenesisTransaction | undefined;
  setGenesis(value?: GenesisTransaction): void;

  hasStateCheckpoint(): boolean;
  clearStateCheckpoint(): void;
  getStateCheckpoint(): StateCheckpointTransaction | undefined;
  setStateCheckpoint(value?: StateCheckpointTransaction): void;

  hasUser(): boolean;
  clearUser(): void;
  getUser(): UserTransaction | undefined;
  setUser(value?: UserTransaction): void;

  getTxnDataCase(): Transaction.TxnDataCase;
  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): Transaction.AsObject;
  static toObject(includeInstance: boolean, msg: Transaction): Transaction.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: Transaction, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): Transaction;
  static deserializeBinaryFromReader(message: Transaction, reader: jspb.BinaryReader): Transaction;
}

export namespace Transaction {
  export type AsObject = {
    timestamp?: aptos_util_timestamp_timestamp_pb.Timestamp.AsObject,
    version: number,
    info?: TransactionInfo.AsObject,
    epoch: number,
    blockHeight: number,
    type: Transaction.TransactionTypeMap[keyof Transaction.TransactionTypeMap],
    blockMetadata?: BlockMetadataTransaction.AsObject,
    genesis?: GenesisTransaction.AsObject,
    stateCheckpoint?: StateCheckpointTransaction.AsObject,
    user?: UserTransaction.AsObject,
  }

  export interface TransactionTypeMap {
    TRANSACTION_TYPE_UNSPECIFIED: 0;
    TRANSACTION_TYPE_GENESIS: 1;
    TRANSACTION_TYPE_BLOCK_METADATA: 2;
    TRANSACTION_TYPE_STATE_CHECKPOINT: 3;
    TRANSACTION_TYPE_USER: 4;
  }

  export const TransactionType: TransactionTypeMap;

  export enum TxnDataCase {
    TXN_DATA_NOT_SET = 0,
    BLOCK_METADATA = 7,
    GENESIS = 8,
    STATE_CHECKPOINT = 9,
    USER = 10,
  }
}

export class BlockMetadataTransaction extends jspb.Message {
  getId(): string;
  setId(value: string): void;

  getRound(): number;
  setRound(value: number): void;

  clearEventsList(): void;
  getEventsList(): Array<Event>;
  setEventsList(value: Array<Event>): void;
  addEvents(value?: Event, index?: number): Event;

  getPreviousBlockVotesBitvec(): Uint8Array | string;
  getPreviousBlockVotesBitvec_asU8(): Uint8Array;
  getPreviousBlockVotesBitvec_asB64(): string;
  setPreviousBlockVotesBitvec(value: Uint8Array | string): void;

  getProposer(): string;
  setProposer(value: string): void;

  clearFailedProposerIndicesList(): void;
  getFailedProposerIndicesList(): Array<number>;
  setFailedProposerIndicesList(value: Array<number>): void;
  addFailedProposerIndices(value: number, index?: number): number;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): BlockMetadataTransaction.AsObject;
  static toObject(includeInstance: boolean, msg: BlockMetadataTransaction): BlockMetadataTransaction.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: BlockMetadataTransaction, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): BlockMetadataTransaction;
  static deserializeBinaryFromReader(message: BlockMetadataTransaction, reader: jspb.BinaryReader): BlockMetadataTransaction;
}

export namespace BlockMetadataTransaction {
  export type AsObject = {
    id: string,
    round: number,
    eventsList: Array<Event.AsObject>,
    previousBlockVotesBitvec: Uint8Array | string,
    proposer: string,
    failedProposerIndicesList: Array<number>,
  }
}

export class GenesisTransaction extends jspb.Message {
  hasPayload(): boolean;
  clearPayload(): void;
  getPayload(): WriteSet | undefined;
  setPayload(value?: WriteSet): void;

  clearEventsList(): void;
  getEventsList(): Array<Event>;
  setEventsList(value: Array<Event>): void;
  addEvents(value?: Event, index?: number): Event;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): GenesisTransaction.AsObject;
  static toObject(includeInstance: boolean, msg: GenesisTransaction): GenesisTransaction.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: GenesisTransaction, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): GenesisTransaction;
  static deserializeBinaryFromReader(message: GenesisTransaction, reader: jspb.BinaryReader): GenesisTransaction;
}

export namespace GenesisTransaction {
  export type AsObject = {
    payload?: WriteSet.AsObject,
    eventsList: Array<Event.AsObject>,
  }
}

export class StateCheckpointTransaction extends jspb.Message {
  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): StateCheckpointTransaction.AsObject;
  static toObject(includeInstance: boolean, msg: StateCheckpointTransaction): StateCheckpointTransaction.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: StateCheckpointTransaction, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): StateCheckpointTransaction;
  static deserializeBinaryFromReader(message: StateCheckpointTransaction, reader: jspb.BinaryReader): StateCheckpointTransaction;
}

export namespace StateCheckpointTransaction {
  export type AsObject = {
  }
}
// ...
export namespace UserTransaction {
  export type AsObject = {
    request?: UserTransactionRequest.AsObject,
    eventsList: Array<Event.AsObject>,
  }
}
// ...
export class TransactionInfo extends jspb.Message {
  getHash(): Uint8Array | string;
  getHash_asU8(): Uint8Array;
  getHash_asB64(): string;
  setHash(value: Uint8Array | string): void;

  getStateChangeHash(): Uint8Array | string;
  getStateChangeHash_asU8(): Uint8Array;
  getStateChangeHash_asB64(): string;
  setStateChangeHash(value: Uint8Array | string): void;

  getEventRootHash(): Uint8Array | string;
  getEventRootHash_asU8(): Uint8Array;
  getEventRootHash_asB64(): string;
  setEventRootHash(value: Uint8Array | string): void;

  hasStateCheckpointHash(): boolean;
  clearStateCheckpointHash(): void;
  getStateCheckpointHash(): Uint8Array | string;
  getStateCheckpointHash_asU8(): Uint8Array;
  getStateCheckpointHash_asB64(): string;
  setStateCheckpointHash(value: Uint8Array | string): void;

  getGasUsed(): number;
  setGasUsed(value: number): void;

  getSuccess(): boolean;
  setSuccess(value: boolean): void;

  getVmStatus(): string;
  setVmStatus(value: string): void;

  getAccumulatorRootHash(): Uint8Array | string;
  getAccumulatorRootHash_asU8(): Uint8Array;
  getAccumulatorRootHash_asB64(): string;
  setAccumulatorRootHash(value: Uint8Array | string): void;

  clearChangesList(): void;
  getChangesList(): Array<WriteSetChange>;
  setChangesList(value: Array<WriteSetChange>): void;
  addChanges(value?: WriteSetChange, index?: number): WriteSetChange;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): TransactionInfo.AsObject;
  static toObject(includeInstance: boolean, msg: TransactionInfo): TransactionInfo.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: TransactionInfo, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): TransactionInfo;
  static deserializeBinaryFromReader(message: TransactionInfo, reader: jspb.BinaryReader): TransactionInfo;
}

export namespace TransactionInfo {
  export type AsObject = {
    hash: Uint8Array | string,
    stateChangeHash: Uint8Array | string,
    eventRootHash: Uint8Array | string,
    stateCheckpointHash: Uint8Array | string,
    gasUsed: number,
    success: boolean,
    vmStatus: string,
    accumulatorRootHash: Uint8Array | string,
    changesList: Array<WriteSetChange.AsObject>,
  }
}
// ...
export class WriteSetChange extends jspb.Message {
  getType(): WriteSetChange.TypeMap[keyof WriteSetChange.TypeMap];
  setType(value: WriteSetChange.TypeMap[keyof WriteSetChange.TypeMap]): void;

  hasDeleteModule(): boolean;
  clearDeleteModule(): void;
  getDeleteModule(): DeleteModule | undefined;
  setDeleteModule(value?: DeleteModule): void;

  hasDeleteResource(): boolean;
  clearDeleteResource(): void;
  getDeleteResource(): DeleteResource | undefined;
  setDeleteResource(value?: DeleteResource): void;

  hasDeleteTableItem(): boolean;
  clearDeleteTableItem(): void;
  getDeleteTableItem(): DeleteTableItem | undefined;
  setDeleteTableItem(value?: DeleteTableItem): void;

  hasWriteModule(): boolean;
  clearWriteModule(): void;
  getWriteModule(): WriteModule | undefined;
  setWriteModule(value?: WriteModule): void;

  hasWriteResource(): boolean;
  clearWriteResource(): void;
  getWriteResource(): WriteResource | undefined;
  setWriteResource(value?: WriteResource): void;

  hasWriteTableItem(): boolean;
  clearWriteTableItem(): void;
  getWriteTableItem(): WriteTableItem | undefined;
  setWriteTableItem(value?: WriteTableItem): void;

  getChangeCase(): WriteSetChange.ChangeCase;
  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): WriteSetChange.AsObject;
  static toObject(includeInstance: boolean, msg: WriteSetChange): WriteSetChange.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: WriteSetChange, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): WriteSetChange;
  static deserializeBinaryFromReader(message: WriteSetChange, reader: jspb.BinaryReader): WriteSetChange;
}

export namespace WriteSetChange {
  export type AsObject = {
    type: WriteSetChange.TypeMap[keyof WriteSetChange.TypeMap],
    deleteModule?: DeleteModule.AsObject,
    deleteResource?: DeleteResource.AsObject,
    deleteTableItem?: DeleteTableItem.AsObject,
    writeModule?: WriteModule.AsObject,
    writeResource?: WriteResource.AsObject,
    writeTableItem?: WriteTableItem.AsObject,
  }

  export interface TypeMap {
    TYPE_UNSPECIFIED: 0;
    TYPE_DELETE_MODULE: 1;
    TYPE_DELETE_RESOURCE: 2;
    TYPE_DELETE_TABLE_ITEM: 3;
    TYPE_WRITE_MODULE: 4;
    TYPE_WRITE_RESOURCE: 5;
    TYPE_WRITE_TABLE_ITEM: 6;
  }

  export const Type: TypeMap;

  export enum ChangeCase {
    CHANGE_NOT_SET = 0,
    DELETE_MODULE = 2,
    DELETE_RESOURCE = 3,
    DELETE_TABLE_ITEM = 4,
    WRITE_MODULE = 5,
    WRITE_RESOURCE = 6,
    WRITE_TABLE_ITEM = 7,
  }
}
// ...
  export const Type: TypeMap;